Fixed quoted words
